change player x and y double instead int

+++ cub.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aamoussa <aamoussa@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ahamdy <ahamdy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/08 18:40:05 by aamoussa          #+#    #+#             */
-/*   Updated: 2023/01/09 17:08:55 by aamoussa         ###   ########.fr       */
+/*   Updated: 2023/01/10 15:06:49 by ahamdy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,8 +49,8 @@ typedef struct s_map_data
 // player struct
 typedef struct s_player
 {
-	int x;
-	int y;
+	double x;
+	double y;
 	// int	radius;
 	// int	turnDirection;
 	// int walkDirection;
